Move SemanticDocumentationArticle to Features\Documentation namespaces
