+++ packages/framework/resources/views/components/docs/documentation-article.blade.php
@@ -1,5 +1,5 @@
 @props([
-    /** @var \Hyde\Support\Models\SemanticDocumentationArticle */
+    /** @var \Hyde\Framework\Features\Documentation\SemanticDocumentationArticle */
     'document'
 ])
